looking at windows build errors for portablexdr with new CI runners

+++ extlib/portablexdr-4.9.1/xdr_mem.c
@@ -49,6 +49,7 @@ static char sccsid[] = "@(#)xdr_mem.c 1.19 87/08/11 Copyr 1984 Sun Micro";
 #endif
 
 #include <string.h>
+#include <stdint.h>
 
 #include "rpc/types.h"
 #include "rpc/xdr.h"
@@ -61,7 +62,7 @@ static bool_t	xdrmem_getbytes();
 static bool_t	xdrmem_putbytes();
 static u_int	xdrmem_getpos();
 static bool_t	xdrmem_setpos();
-static long *	xdrmem_inline();
+static int32_t* xdrmem_inline();
 static void	xdrmem_destroy();
 
 static struct	xdr_ops xdrmem_ops = {
@@ -169,7 +170,7 @@ xdrmem_setpos(xdrs, pos)
 	register caddr_t newaddr = xdrs->x_base + pos;
 	register caddr_t lastaddr = xdrs->x_private + xdrs->x_handy;
 
-	if ((long)newaddr > (long)lastaddr)
+	if ((intptr_t)newaddr > (intptr_t)lastaddr)
 		return (FALSE);
 	xdrs->x_private = newaddr;
 	xdrs->x_handy = lastaddr - newaddr;
